refactor: reorganize configuration imports and move constants to config.py
